fix: extract BoardEmptyState into own component
